Split .models to 4 submodules
- Separate common code and code specific to MACRO, MESSAGE, and
  MESSAGE_MACRO.
- Add .models.__getattr__() for deprecated imports with warnings.
